Refactor Kubernetes Service Account to genai/common
This PR resolves a workflow issue for me iterating on this repo: the
`sed` to template the project for the KSA->SA mapping is kind of
obnoxious from a `git` perspective.

Instead moves common KSA to a new module in common and introduces a
"common" module that pulls all the common deps.

This doesn't make the files ignored since they're still tracked, but
it does mean you can just avoid adding them to commits.